GeminiCall/pyproject.toml
[project]
name = "geminicall"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.3",
    "google-genai>=1.62.0",
    "openai>=1.0.0",
    "httpx>=0.28.1",
    "mcp>=1.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "uvicorn>=0.40.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
